requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.4
protobuf>=4.21.0
supabase==2.4.2
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1